adding test fixtures for spotlight_bloc tests
